requirements.txt
fastapi[standard]
SQLAlchemy
asyncpg
python-dotenv
pydantic_settings
alembic
psycopg2-binary
passlib[bcrypt]
pyjwt
redis
argon2_cffi